requirements.txt
alembic==1.9.2
APScheduler==3.9.1.post1
click==8.1.3
colorama==0.4.6
flake8==6.0.0
Flask==2.2.2
Flask-APScheduler==1.12.4
Flask-Migrate==4.0.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
greenlet==2.0.1
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
numpy==1.24.1
pandas==1.5.3
pycodestyle==2.10.0
pyflakes==3.0.1
pyodbc==4.0.35
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.46
tomli==2.0.1
types-python-dateutil==2.8.19.6
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
Werkzeug==2.2.2
WTForms==3.0.1
